Refactor formation files
